fix(new-client): num filter should handle float values (734)

+++ src/new-client/src/App/Trades/TradesGrid/NumFilter.tsx
@@ -37,7 +37,7 @@ const FilterValueInput: React.FC<{
     onChange={({ target: { value } }) => {
       onColFilterEnterNum(field, {
         ...selected,
-        [fieldValueName]: value ? parseInt(value) : null,
+        [fieldValueName]: value ? parseFloat(value) : null,
       })
     }}
   />
